Rename esmvaltool.models to esmvaltool.schema
When we talk about models we mean hydrological models.
