+++ src/ewatercycle/plugins/wflow/forcing.py
@@ -5,7 +5,7 @@ from typing import Dict, Optional
 
 from ewatercycle.base.forcing import DefaultForcing
 from ewatercycle.esmvaltool.builder import RecipeBuilder
-from ewatercycle.esmvaltool.models import Dataset
+from ewatercycle.esmvaltool.schema import Dataset
 
 
 class WflowForcing(DefaultForcing):
